fix: break calibration import cycles
